clean up parser
The parser was cleaned up and prepared for some future tasks.
Mainly memory management is better.

+++ src/utility/xalloc.c
@@ -121,7 +121,7 @@ char *xasprintf(const char *format, ...)
 
     va_start(list, format);
     /* +1 for the null terminator */
-    result = xmalloc(total_size + 1);
+    result = xmalloc((size_t) total_size + 1);
     (void) vsprintf(result, format, list);
     va_end(list);
